requirements.txt
# Core Dependencies
streamlit==1.31.0
fastapi==0.109.0
uvicorn==0.27.0

# AI Integration
google-generativeai==0.3.2

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3

# Optional but recommended
requests==2.31.0